完成可选徽章预览接口 完成佩戴徽章接口

+++ chatroom-chat-server/src/main/java/com/dazhou/chatroom/common/user/domain/entity/UserBackpack.java
@@ -42,7 +42,7 @@ public class UserBackpack implements Serializable {
      * 物品id
      */
     @TableField("item_id")
-    private Integer itemId;
+    private Long itemId;
 
     /**
      * 使用状态 0.待使用 1已使用
